--- a/.gitignore
+++ b/.gitignore
@@ -8,7 +8,7 @@
 **/*.tar.gz
 **/playground
 **/wandb
-
+**/tensorboard_log
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]