… recipe/grpo/main_grpo.py
… recipe/moe/moe_trainer/config/ppo_megatron_trainer.yaml
… recipe/moe/moe_trainer/config/ppo_trainer.yaml
… recipe/moe/moe_trainer/config/sft_trainer.yaml
… recipe/moe/moe_trainer/debug_data/sft.json
… recipe/moe/moe_trainer/fsdp_sft_trainer.py
… recipe/moe/moe_trainer/main_ppo.py
… recipe/moe/moe_trainer/modeling_xdgmoe.py
… recipe/moe/moe_trainer/ppo/__init__.py
… recipe/moe/moe_trainer/ppo/core_algos.py
… recipe/moe/moe_trainer/ppo/metric_utils.py
… recipe/moe/moe_trainer/ppo/ray_trainer.py
… recipe/moe/moe_trainer/ppo/reward.py
… recipe/moe/moe_trainer/runtime_env.yaml
… recipe/moe/scripts/run_xdg3_debug.sh
… recipe/moe/scripts/run_xdg3_grpo_debug.sh
… recipe/moe/scripts/run_xdg_debug.sh
… recipe/moe/test/inputs.pkl
… recipe/moe/test/test_check_dump.ipynb
… recipe/moe/test/test_forward_check.py
… verl/__init__.py
… verl/models/dots/sglang.py
… verl/models/dots/vllm.py
… verl/models/mcore/config_converter.py
… verl/models/mcore/hf_mcore_loader.py
… verl/models/mcore/model_forward_fused.py
… verl/models/mcore/model_initializer.py
… verl/models/mcore/qwen2_5_vl/attention.py
… verl/models/mcore/qwen2_5_vl/vision_transformer_block.py
… verl/models/mcore/registry.py